requirements.txt
requests
BeautifulSoup4
streamlit 
psycopg2-binary
python-dotenv